chore(routing): extract update of routing rule statuses in a concern
