+++ app/models/procedure.rb
@@ -12,6 +12,7 @@ class Procedure < ApplicationRecord
   include ProcedureCloneConcern
   include PiecesJointesListConcern
   include ColumnsConcern
+  include RoutingRuleStatusesConcern
 
   include Discard::Model
   self.discard_column = :hidden_at
@@ -726,43 +727,6 @@ class Procedure < ApplicationRecord
       gi = groupe_instructeurs.create(label: GroupeInstructeur::DEFAUT_LABEL)
       self.update(defaut_groupe_instructeur_id: gi.id)
     end
-  end
-
-  def update_all_groupes_rule_validity_status
-    valid_ids = []
-    invalid_ids = []
-
-    self.groupe_instructeurs.includes(:procedure).find_each do |gi|
-      if gi.valid_rule?
-        valid_ids << gi.id
-      else
-        invalid_ids << gi.id
-      end
-    end
-
-    GroupeInstructeur.where(id: valid_ids).update_all(valid_routing_rule: true) if valid_ids.any?
-    GroupeInstructeur.where(id: invalid_ids).update_all(valid_routing_rule: false) if invalid_ids.any?
-  end
-
-  def update_all_groupes_rule_unicity_status
-    # Get ids from groupe_instructeurs with same routing rule
-    rule_gis = Hash.new { |h, k| h[k] = [] }
-
-    self.groupe_instructeurs.filter { it.routing_rule.present? }.each_with_object(rule_gis) do |gi, h|
-      h[gi.routing_rule] << gi.id
-    end
-
-    duplicate_ids = rule_gis.filter { |_k, gi_ids| gi_ids.size > 1 }.map(&:second).flatten
-
-    # Update unique_routing_rule in all groups
-    self.groupe_instructeurs.update_all(unique_routing_rule: true)
-    self.groupe_instructeurs.where(routing_rule: nil).update_all(unique_routing_rule: false)
-    self.groupe_instructeurs.where(id: duplicate_ids).update_all(unique_routing_rule: false)
-  end
-
-  def update_all_groupes_rule_statuses
-    update_all_groupes_rule_validity_status
-    update_all_groupes_rule_unicity_status
   end
 
   def create_generic_labels
